+++ bukkit/changelog.md
@@ -1,1 +1,2 @@
 - Improved handling of invalid bind addresses
+- Fixed compatibility with Folia